# api/openapi/ops-platform.yaml
openapi: 3.0.3
info:
  title: Ops Platform API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post:
      summary: Login with local email credentials
      description: When auth.mfa_enabled is true, password validation is followed by mandatory TOTP enrollment or verification. When false, a valid password issues the JWT token pair without a TOTP challenge, even if the user already has an MFA binding.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "429":
          description: Too many failed login attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseEnvelope"
  /auth/refresh:
    post:
      summary: Refresh JWT token pair
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
  /auth/mfa/setup:
    post:
      summary: Start mandatory MFA enrollment from a login challenge
      operationId: setupLoginMFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFASetupRequest"
      responses:
        "200":
          description: TOTP enrollment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFASetupEnvelope"
  /auth/mfa/verify:
    post:
      summary: Verify a login TOTP challenge and issue a JWT token pair
      operationId: verifyLoginMFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAVerifyRequest"
      responses:
        "200":
          description: MFA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "429":
          description: Too many failed MFA verification attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseEnvelope"
  /auth/profile:
    get:
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalEnvelope"
  /auth/mfa/status:
    get:
      summary: Current user's MFA status and server policy
      operationId: getMFAStatus
      security:
        - bearerAuth: []
      responses:
        "200":
          description: MFA status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFAStatusEnvelope"
  /auth/mfa/enrollment:
    post:
      summary: Start MFA enrollment for the current authenticated user
      operationId: startMFAEnrollment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: TOTP enrollment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFASetupEnvelope"
  /auth/mfa/enable:
    post:
      summary: Verify and enable MFA for the current user
      operationId: enableMFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFAVerifyRequest"
      responses:
        "200":
          description: MFA enabled and replacement token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
  /auth/mfa/disable:
    post:
      summary: Disable MFA for the current user when server policy permits it
      operationId: disableMFA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFADisableRequest"
      responses:
        "200":
          description: MFA disabled and replacement token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
...
  /roles:
    get:
      summary: List platform roles
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListEnvelope"
  /users:
    get:
      summary: List platform users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListEnvelope"
    post:
      summary: Create local platform user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
  /users/{id}/status:
    put:
      summary: Update platform user status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: User status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
  /users/{id}/roles:
    put:
      summary: Replace platform user roles
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRolesRequest"
      responses:
        "200":
          description: User roles updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
  /users/{id}/mfa:
    delete:
      summary: Reset another user's MFA enrollment
      operationId: resetUserMFA
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: User MFA reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
...
  /namespaces/{namespace}/pods/{pod}/logs:
    get:
      summary: Query retained pod logs from Kubernetes API
      description: Reads logs still retained by the node for the current container instance or its immediately previous instance. This is not durable cross-Pod history; P1 Loki provides long-term querying.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
        - name: container
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: previous
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: from
          in: query
          description: Return log entries at or after this RFC3339 timestamp.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Log result
  /namespaces/{namespace}/pods/{pod}/logs/follow:
    get:
      summary: Follow pod logs over WebSocket
      description: Connect to /ws/v1 and stream Kubernetes follow=true logs. The client offers subprotocols ops-platform.logs.v1 and bearer.<access-token>; the server negotiates ops-platform.logs.v1. Previous container logs cannot be followed.
      servers:
        - url: /ws/v1
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
        - name: container
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: limit
          in: query
          description: Number of retained lines sent before following new lines.
          schema:
            type: integer
            default: 200
            maximum: 2000
        - name: from
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "101":
          description: WebSocket stream established; messages use LogStreamMessage.
...
  /logs:
    get:
      summary: Query logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Log result
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
    Pod:
      name: pod
      in: path
      required: true
      schema:
        type: string
...
    UserID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
    ResponseEnvelope:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    MFASetupRequest:
      type: object
      required:
        - mfa_token
      properties:
        mfa_token:
          type: string
    MFAVerifyRequest:
      type: object
      required:
        - mfa_token
        - code
      properties:
        mfa_token:
          type: string
        code:
          type: string
          pattern: "^[0-9]{6}$"
    MFADisableRequest:
      type: object
      required:
        - password
        - code
      properties:
        password:
          type: string
          format: password
        code:
          type: string
          pattern: "^[0-9]{6}$"
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        provider:
          type: string
        status:
          type: string
          enum:
            - active
            - disabled
        created_at:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
        mfa_enabled:
          type: boolean
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        roles:
          type: array
          items:
            type: string
    UpdateUserStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - active
            - disabled
    UpdateUserRolesRequest:
      type: object
      required:
        - roles
      properties:
        roles:
          type: array
          minItems: 1
          items:
            type: string
    Principal:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        mfa_enabled:
          type: boolean
        mfa_verified:
          type: boolean
    LoginResult:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: "#/components/schemas/Principal"
        mfa_required:
          type: boolean
        mfa_setup_required:
          type: boolean
        mfa_token:
          type: string
    MFASetupResult:
      type: object
      required:
        - secret
        - provisioning_uri
        - mfa_token
        - expires_in
      properties:
        secret:
          type: string
        provisioning_uri:
          type: string
        mfa_token:
          type: string
        expires_in:
          type: integer
          format: int64
    MFAStatus:
      type: object
      required:
        - enabled
        - required
      properties:
        enabled:
          type: boolean
          description: Whether the current user has a stored TOTP binding
        required:
          type: boolean
          description: Whether auth.mfa_enabled currently requires TOTP during login
    LoginEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LoginResult"
    PrincipalEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Principal"
    MFASetupEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MFASetupResult"
    MFAStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MFAStatus"
    UserEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/User"
    UserListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/User"
    RoleListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Role"